refactor (Lotto) : Lotto를 리팩토링 한다
- 정적 팩토리 메서드를 사용
- lotto 패키지로 분리
- 상속 가능하도록 private 접근 제어자를 protected로 변경
