+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.enums.LottoPrize;
+import lotto.domain.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,21 +14,21 @@ class LottoTest {
     @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
     @Test
     void createLottoByOverSize() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
+        assertThatThrownBy(() -> Lotto.from(List.of(1, 2, 3, 4, 5, 6, 7)))
             .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
     @Test
     void createLottoByDuplicatedNumber() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
+        assertThatThrownBy(() -> Lotto.from(List.of(1, 2, 3, 4, 5, 5)))
             .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("로또 번호가 null 이라면 예외가 발생한다.")
     @Test
     void createLottoByNull() {
-        assertThatThrownBy(() -> new Lotto(null))
+        assertThatThrownBy(() -> Lotto.from(null))
             .isInstanceOf(IllegalArgumentException.class);
     }
 }
